fix: created Section Component using Toggler component and used in lead/deal
